refactor: jounal page index.tsx로 파일명 변경
