+++ frontend/src/pages/Journals/index.tsx
